dashboards/ops/requirements.txt
streamlit>=1.30
psycopg2-binary>=2.9
pandas>=1.5
